legend-query: create QueryEditorStore and simplify LegendQueryStore
