+++ .changeset/light-bananas-agree.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-query': major
+---
+
+**BREAKING CHANGE:** Remove logic from `LegendQueryStore` to separate between query setup and edit.
