+++ .changeset/tame-apples-play.md
@@ -1,0 +1,19 @@
+---
+'@finos/eslint-plugin-legend-studio': patch
+'@finos/legend-application': patch
+'@finos/legend-dev-utils': patch
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-extension-mapping-generation': patch
+'@finos/legend-graph': patch
+'@finos/legend-query': patch
+'@finos/legend-query-deployment': patch
+'@finos/legend-server-depot': patch
+'@finos/legend-server-sdlc': patch
+'@finos/legend-shared': patch
+'@finos/legend-studio': patch
+'@finos/legend-studio-deployment': patch
+'@finos/legend-studio-extension-query-builder': patch
+'@finos/legend-taxonomy': patch
+'@finos/legend-taxonomy-deployment': patch
+'@finos/legend-tracer-extension-zipkin': patch
+---
