+++ packages/legend-query/src/components/QuerySetupStoreProvider.tsx
@@ -16,9 +16,12 @@
 
 import { createContext, useContext } from 'react';
 import { useLocalObservable } from 'mobx-react-lite';
-import { QuerySetupStore } from '../stores/LegendQuerySetupStore.js';
+import { QuerySetupStore } from '../stores/QuerySetupStore.js';
 import { useLegendQueryStore } from './LegendQueryStoreProvider.js';
 import { guaranteeNonNullable } from '@finos/legend-shared';
+import { useDepotServerClient } from '@finos/legend-server-depot';
+import { useApplicationStore } from '@finos/legend-application';
+import type { LegendQueryConfig } from '../application/LegendQueryConfig.js';
 
 const QuerySetupStoreContext = createContext<QuerySetupStore | undefined>(
   undefined,
@@ -27,8 +30,17 @@ const QuerySetupStoreContext = createContext<QuerySetupStore | undefined>(
 export const QuerySetupStoreProvider: React.FC<{
   children: React.ReactNode;
 }> = ({ children }) => {
+  const applicationStore = useApplicationStore<LegendQueryConfig>();
+  const depotServerClient = useDepotServerClient();
   const queryStore = useLegendQueryStore();
-  const store = useLocalObservable(() => new QuerySetupStore(queryStore));
+  const store = useLocalObservable(
+    () =>
+      new QuerySetupStore(
+        applicationStore,
+        depotServerClient,
+        queryStore.pluginManager,
+      ),
+  );
   return (
     <QuerySetupStoreContext.Provider value={store}>
       {children}
@@ -39,5 +51,5 @@ export const QuerySetupStoreProvider: React.FC<{
 export const useQuerySetupStore = (): QuerySetupStore =>
   guaranteeNonNullable(
     useContext(QuerySetupStoreContext),
-    `Can't find Query setup store in context`,
+    `Can't find query setup store in context`,
   );
